refactor(sequencer): move AutomineSequencer to its own subdirectory + add README
Pure reorganization. AutomineSequencer + factory now live under
src/sequencer/automine/ with a dedicated README explaining the design
and when to use AUTOMINE_E2E_OPTS. No behavior change.
