+++ yarn-project/sequencer-client/src/sequencer/index.ts
@@ -1,5 +1,5 @@
-export * from './automine_factory.js';
-export * from './automine_sequencer.js';
+export * from './automine/automine_factory.js';
+export * from './automine/automine_sequencer.js';
 export * from './checkpoint_proposal_job.js';
 export * from './checkpoint_voter.js';
 export * from './config.js';
